01_speechgraphics_aws/python/requirements.txt
boto3
pyjwt
Django
warrant
pynamodb
gunicorn

# development dependencies
tox
black
flake8
